# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface RoomType {
  id: string;
  name: string;
  description: string;
}

export interface StylePreset {
  id: string;
  name: string;
  tagline: string;
  description: string;
  accentColor: string;
  /** Relative URL path to the style preview image (served by the API). */
  previewImage: string;
}

export interface Product {
  id: string;
  name: string;
  category: string;
  color: string;
  price: number;
  currency: string;
  /** Room type ids this product is eligible for. */
  roomTypes: string[];
  /** Functional role in the room (e.g. sofa, bed, rug), used for de-cluttered selection. */
  role: string;
  /** Relative URL path to the product image (served by the API). */
  imageUrl: string;
  /** Link to purchase the product on IKEA. */
  buyUrl: string;
}

export interface RedesignRequest {
  /** Base64-encoded room photo (no data URI prefix). */
  image: string;
  styleId: string;
  roomTypeId: string;
  /** Anonymous device identifier the redesign is saved under. */
  deviceId: string;
  /** Optional subset of the room's auto-selected product ids to include. When omitted or empty, the server's default de-cluttered selection is used. */
  productIds?: string[];
}

export interface RegenerateRequest {
  /** Anonymous device identifier that owns the redesign. */
  deviceId: string;
  /** The curated set of product ids to ground the regenerated room in. */
  productIds: string[];
}

export interface Redesign {
  id: string;
  /** Creation time as epoch milliseconds. */
  createdAt: number;
  deviceId: string;
  styleId: string;
  styleName: string;
  roomTypeId: string;
  roomName: string;
  /** Base64-encoded original room photo. */
  originalImage: string;
  /** Base64-encoded redesigned room image (PNG). */
  redesignedImage: string;
  products: Product[];
}

export type ListProductsParams = {
roomTypeId?: string;
};

export type ListRedesignsParams = {
deviceId: string;
};