# lib/api-zod/src/generated/types/redesign.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface Redesign {
  id: string;
  /** Creation time as epoch milliseconds. */
  createdAt: number;
  deviceId: string;
  styleId: string;
  styleName: string;
  roomTypeId: string;
  roomName: string;
  /** Base64-encoded original room photo. */
  originalImage: string;
  /** Base64-encoded redesigned room image (PNG). */
  redesignedImage: string;
  products: Product[];
}