# lib/api-zod/src/generated/types/regenerateRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegenerateRequest {
  /** Anonymous device identifier that owns the redesign. */
  deviceId: string;
  /** The curated set of product ids to ground the regenerated room in. */
  productIds: string[];
}